# Better handle String hashCode collisions in JsLookup