--- a/benchmarks/src/main/scala/play/api/libs/json/JsonParsing_01_ParseManyFields.scala
+++ b/benchmarks/src/main/scala/play/api/libs/json/JsonParsing_01_ParseManyFields.scala
@@ -6,12 +6,20 @@ package play.api.libs.json
 
 import org.openjdk.jmh.annotations._
 
+/**
+  * @see https://github.com/playframework/play-json/pull/193
+  */
 @State(Scope.Benchmark)
 class JsonParsing_01_ParseManyFields {
   @Param(Array("10", "100", "1000", "10000", "100000"))
   var n: Int = 100
 
   var stringToParse: String = _
+
+  case class Example(s: String)
+  object Example {
+    implicit val reads = Json.reads[Example]
+  }
 
   @Setup
   def setup(): Unit = {
@@ -21,7 +29,22 @@ class JsonParsing_01_ParseManyFields {
   }
 
   @Benchmark
-  def parseObject(): Unit = {
+  def parseObject(): JsValue = {
     Json.parse(stringToParse)
+  }
+
+  @Benchmark
+  def parseObjectAs(): Example = {
+    Json.parse(stringToParse).as[Example]
+  }
+
+  @Benchmark
+  def parseObjectLookup(): JsValue = {
+    (Json.parse(stringToParse) \ "s").as[JsString]
+  }
+
+  @Benchmark
+  def parseObjectValue(): Option[JsValue] = {
+    Json.parse(stringToParse).as[JsObject].value.get("s")
   }
 }