--- a/play-json/shared/src/main/scala/play/api/libs/json/JsLookup.scala
+++ b/play-json/shared/src/main/scala/play/api/libs/json/JsLookup.scala
@@ -68,7 +68,7 @@ case class JsLookup(result: JsLookupResult) extends AnyVal {
     case JsDefined(x) =>
       x match {
         case arr: JsObject =>
-          arr.value.get(fieldName) match {
+          arr.underlying.get(fieldName) match {
             case Some(x) => x
             case None    => throw new NoSuchElementException(String.valueOf(fieldName))
           }
@@ -101,7 +101,7 @@ case class JsLookup(result: JsLookupResult) extends AnyVal {
    */
   def \(fieldName: String): JsLookupResult = result match {
     case JsDefined(obj @ JsObject(_)) =>
-      obj.value
+      obj.underlying
         .get(fieldName)
         .map(JsDefined.apply)
         .getOrElse(JsUndefined(s"'$fieldName' is undefined on object: $obj"))
@@ -117,7 +117,7 @@ case class JsLookup(result: JsLookupResult) extends AnyVal {
    */
   def \\(fieldName: String): collection.Seq[JsValue] = result match {
     case JsDefined(obj: JsObject) =>
-      obj.value.foldLeft(Seq[JsValue]())((o, pair) =>
+      obj.underlying.foldLeft(Seq[JsValue]())((o, pair) =>
         pair match {
           case (key, value) if key == fieldName => o ++ (value +: (value \\ fieldName))
           case (_, value)                       => o ++ (value \\ fieldName)